Improves API to support up to 2^63 key-value pairs.

+++ src/ZoneTree/Collections/BTree/BTreeSeekableIterator.cs
@@ -85,7 +85,7 @@ public class BTreeSeekableIterator<TKey, TValue>
         return HasCurrent;
     }
 
-    public void Skip(int offset)
+    public void Skip(long offset)
     {
         throw new NotSupportedException();
     }
