+++ src/ZoneTree/Collections/BTree/FrozenBTreeSeekableIterator.cs
@@ -85,7 +85,7 @@ public class FrozenBTreeSeekableIterator<TKey, TValue>
         return HasCurrent;
     }
 
-    public void Skip(int offset)
+    public void Skip(long offset)
     {
         throw new NotSupportedException();
     }
